refactor: Mudança das configurações do swagger
