+++ src/main/java/tech/ada/java/agendamentoconsultas/security/service/AuthorizationService.java
@@ -1,16 +1,10 @@
 package tech.ada.java.agendamentoconsultas.security.service;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
 import lombok.RequiredArgsConstructor;
 import tech.ada.java.agendamentoconsultas.repository.PatientRepository;
 
@@ -23,24 +17,5 @@ public class AuthorizationService implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return patientRepository.findByEmail(username);
-    }
-    
-    @Bean
-    public OpenAPI customOpenAPI() {
-        final String securitySchemeName = "basicAuth";
-        final String apiTitle = "To-Do List API";
-        return new OpenAPI()
-            .addSecurityItem(new SecurityRequirement().addList(securitySchemeName))
-            .components(
-                new Components()
-                    .addSecuritySchemes(securitySchemeName,
-                        new SecurityScheme()
-                            .name(securitySchemeName)
-                            .type(SecurityScheme.Type.HTTP)
-                            .scheme("basic")
-//                            .bearerFormat("JWT")
-                    )
-            )
-            .info(new Info().title(apiTitle).version("0.0.1-SNAPSHOT"));
     }
 }
